Estimate the proportion of asymptomatic cases by capture/recapture

Estimate the proportion of asymptomatic cases by capture/recapture

estimate_asympto(date, cases, deaths, bounds = c("lower", "upper"))

Arguments

date

A vector containing the dates

cases

A numeric vector containing the number of new cases at each date (not the cumulative number of cases).

deaths

A numeric vector containing the number of new deaths at each date (not the cumulative number of cases).

bounds

"lower", "upper", or both c("lower, "upper") (the default), telling which bounds of the number of asymptomatic cases are computed. associated with the lower and upper bounds. estimation. FALSE is never the recommended setting excepted for testing or debugging purpose.

Value

A data.frame with two or three columns (depending on the value of the bounds argument):

date: the original date column

lower: the lower bound of asymptomatic cases

upper: the upper bound of asymptomatic cases

Note

Please note there is currently some discussion about the validity of the methods implemented in this package. You should read carefully the original articles, alongside this answer from Li et al. (2022) before using this package in your project.

References

Böhning D., Rocchetti I., Maruotti A., Holling H. (2020), Estimating the undetected infections in the Covid-19 outbreak by harnessing capture–recapture methods, International Journal of Infectious Diseases, 97, p197-201, doi:10.1016/j.ijid.2020.06.009 .

Rocchetti I., Böhning D., Holling H., Maruotti A., (2020), Estimating the size of undetected cases of the SARS-CoV-2 outbreak in Europe: An upperbound estimator, medRxiv, doi:10.1101/2020.07.14.20153445 .

Li M., Dushoff J., Earn D.J.D., Bolker B.M. (2022), Evaluating undercounts in epidemics: responsi to Maruotti et al. 2022, arXiv, doi:10.48550/ARXIV.2209.11334 .

Examples

d <- readRDS(system.file("extdata", "covid19_italy.rds", package = "asymptor"))
head(d)
#>         date new_cases new_deaths
#> 1 2020-01-02         0          0
#> 2 2020-01-03         0          0
#> 3 2020-01-04         0          0
#> 4 2020-01-05         0          0
#> 5 2020-01-06         0          0
#> 6 2020-01-07         0          0

estimate_asympto(d$date, d$new_cases, d$new_deaths)
#>           date lower  upper
#> 1   2020-01-02    NA     NA
#> 2   2020-01-03     0     NA
#> 3   2020-01-04     0     NA
#> 4   2020-01-05     0     NA
#> 5   2020-01-06     0     NA
#> 6   2020-01-07     0     NA
#> 7   2020-01-08     0     NA
#> 8   2020-01-09     0     NA
#> 9   2020-01-10     0     NA
#> 10  2020-01-11     0     NA
#> 11  2020-01-12     0     NA
#> 12  2020-01-13     0     NA
#> 13  2020-01-14     0     NA
#> 14  2020-01-15     0     NA
#> 15  2020-01-16     0     NA
#> 16  2020-01-17     0     NA
#> 17  2020-01-18     0     NA
#> 18  2020-01-19     0     NA
#> 19  2020-01-20     0     NA
#> 20  2020-01-21     0     NA
#> 21  2020-01-22     0     NA
#> 22  2020-01-23     0     NA
#> 23  2020-01-24     0     NA
#> 24  2020-01-25     0     NA
#> 25  2020-01-26     0     NA
#> 26  2020-01-27     0     NA
#> 27  2020-01-28     2     NA
#> 28  2020-01-29     0      2
#> 29  2020-01-30     0     NA
#> 30  2020-01-31     0      1
#> 31  2020-02-01     0     NA
#> 32  2020-02-02     0     NA
#> 33  2020-02-03     0     NA
#> 34  2020-02-04     0     NA
#> 35  2020-02-05     0     NA
#> 36  2020-02-06     0     NA
#> 37  2020-02-07     0     NA
#> 38  2020-02-08     0     NA
#> 39  2020-02-09     0     NA
#> 40  2020-02-10     0     NA
#> 41  2020-02-11     0     NA
#> 42  2020-02-12     0     NA
#> 43  2020-02-13     0     NA
#> 44  2020-02-14     0     NA
#> 45  2020-02-15     0     NA
#> 46  2020-02-16     0     NA
#> 47  2020-02-17     0     NA
#> 48  2020-02-18     0     NA
#> 49  2020-02-19     0     NA
#> 50  2020-02-20     0     NA
#> 51  2020-02-21   182     NA
#> 52  2020-02-22   291    357
#> 53  2020-02-23    44    163
#> 54  2020-02-24   972   1451
#> 55  2020-02-25    39    356
#> 56  2020-02-26    65    318
#> 57  2020-02-27   841   1388
#> 58  2020-02-28   228    745
#> 59  2020-02-29   248    844
#> 60  2020-03-01  1355   2537
#> 61  2020-03-02   212   1154
#> 62  2020-03-03   686   1924
#> 63  2020-03-04   784   2028
#> 64  2020-03-05  1080   2721
#> 65  2020-03-06   838   2615
#> 66  2020-03-07  2091   4799
#> 67  2020-03-08  1995   5067
#> 68  2020-03-09  2312   6240
#> 69  2020-03-10   585   3466
#> 70  2020-03-11  6838  13100
#> 71  2020-03-12  3306   8491
#> 72  2020-03-13  2700   8670
#> 73  2020-03-14  5152  12975
#> 74  2020-03-15  4116  12063
#> 75  2020-03-16  3223  11081
#> 76  2020-03-17  4302  12659
#> 77  2020-03-18  5798  14983
#> 78  2020-03-19  7490  18751
#> 79  2020-03-20  7629  20562
#> 80  2020-03-21  8276  22775
#> 81  2020-03-22  5232  18771
#> 82  2020-03-23  4623  16887
#> 83  2020-03-24  6807  19189
#> 84  2020-03-25  5942  17763
#> 85  2020-03-26  8321  22058
#> 86  2020-03-27  6847  20190
#> 87  2020-03-28  7037  20817
#> 88  2020-03-29  5214  17805
#> 89  2020-03-30  3722  14116
#> 90  2020-03-31  5110  14885
#> 91  2020-04-01  6872  17087
#> 92  2020-04-02  5415  15704
#> 93  2020-04-03  5385  15938
#> 94  2020-04-04  5911  16806
#> 95  2020-04-05  4350  14656
#> 96  2020-04-06  3518  12558
#> 97  2020-04-07  3082  10687
#> 98  2020-04-08  5889  14503
#> 99  2020-04-09  5475  14296
#> 100 2020-04-10  4293  13382
#> 101 2020-04-11  6609  17118
#> 102 2020-04-12  3926  13644
#> 103 2020-04-13  2818  11056
#> 104 2020-04-14  3460  10815
#> 105 2020-04-15  2969   9187
#> 106 2020-04-16  6687  14723
#> 107 2020-04-17  3797  11389
#> 108 2020-04-18  4045  12245
#> 109 2020-04-19  3034  10420
#> 110 2020-04-20  1961   7863
#> 111 2020-04-21  4321  10771
#> 112 2020-04-22  4951  11737
#> 113 2020-04-23  2408   8752
#> 114 2020-04-24  4097  11285
#> 115 2020-04-25  2130   7870
#> 116 2020-04-26  2573   8360
#> 117 2020-04-27  1517   5965
#> 118 2020-04-28  3218   8168
#> 119 2020-04-29  2459   6972
#> 120 2020-04-30  1944   6401
#> 121 2020-05-01  2406   7007
#> 122 2020-05-02  2418   6614
#> 123 2020-05-03  1116   4703
#> 124 2020-05-04  1247   4452
#> 125 2020-05-05  1171   3775
#> 126 2020-05-06  2947   5942
#> 127 2020-05-07  1675   4527
#> 128 2020-05-08  1518   4570
#> 129 2020-05-09  1033   3699
#> 130 2020-05-10   699   2812
#> 131 2020-05-11   886   2732
#> 132 2020-05-12  3428   6185
#> 133 2020-05-13   652   2864
#> 134 2020-05-14  1568   4102
#> 135 2020-05-15   828   2573
#> 136 2020-05-16  1201   3186
#> 137 2020-05-17   622   2245
#> 138 2020-05-18   352   1595
#> 139 2020-05-19  2276   4234
#> 140 2020-05-20   676   2059
#> 141 2020-05-21   807   2306
#> 142 2020-05-22   827   2261
#> 143 2020-05-23   837   2312
#> 144 2020-05-24   454   1795
#> 145 2020-05-25   204   1088
#> 146 2020-05-26   705   1735
#> 147 2020-05-27  1212   2223
#> 148 2020-05-28   682   1918
#> 149 2020-05-29   524   1765
#> 150 2020-05-30   425   1451
#> 151 2020-05-31   367   1238
#> 152 2020-06-01   106    643
#> 153 2020-06-02   813   1696
#> 154 2020-06-03   414   1020
#> 155 2020-06-04   133    595
#> 156 2020-06-05  2880   4188
#> 157 2020-06-06   162    883
#> 158 2020-06-07   177    759
#> 159 2020-06-08   587   1216
#> 160 2020-06-09   395    922
#> 161 2020-06-10   191    664
#> 162 2020-06-11   955   1771
#> 163 2020-06-12    82    583
#> 164 2020-06-13  1388   2608
#> 165 2020-06-14   376   1040
#> 166 2020-06-15   292   1039
#> 167 2020-06-16   163    736
#> 168 2020-06-17   642   1431
#> 169 2020-06-18   419   1086
#> 170 2020-06-19   219    844
#> 171 2020-06-20   337    973
#> 172 2020-06-21   209    752
#> 173 2020-06-22   234    773
#> 174 2020-06-23    73    437
#> 175 2020-06-24   233    733
#> 176 2020-06-25   556   1155
#> 177 2020-06-26   250    835
#> 178 2020-06-27   121    620
#> 179 2020-06-28   195    659
#> 180 2020-06-29    93    434
#> 181 2020-06-30   193    551
#> 182 2020-07-01   285    669
#> 183 2020-07-02   254    677
#> 184 2020-07-03   265    766
#> 185 2020-07-04   271    810
#> 186 2020-07-05   160    654
#> 187 2020-07-06   233    755
#> 188 2020-07-07   106    473
#> 189 2020-07-08   299    769
#> 190 2020-07-09   287    775
#> 191 2020-07-10   348    961
#> 192 2020-07-11   130    647
#> 193 2020-07-12   303    907
#> 194 2020-07-13   128    572
#> 195 2020-07-14    84    415
#> 196 2020-07-15   259    659
#> 197 2020-07-16   366    818
#> 198 2020-07-17   246    769
#> 199 2020-07-18   281    872
#> 200 2020-07-19   193    748
#> 201 2020-07-20   173    672
#> 202 2020-07-21    94    457
#> 203 2020-07-22   655   1327
#> 204 2020-07-23   342    985
#> 205 2020-07-24   209    864
#> 206 2020-07-25   304    998
#> 207 2020-07-26   239    874
#> 208 2020-07-27   114    607
#> 209 2020-07-28   280    831
#> 210 2020-07-29   402   1021
#> 211 2020-07-30   518   1346
#> 212 2020-07-31   379   1274
#> 213 2020-08-01   231   1029
#> 214 2020-08-02   198    861
#> 215 2020-08-03   110    575
#> 216 2020-08-04   232    725
#> 217 2020-08-05   813   1584
#> 218 2020-08-06   425   1300
#> 219 2020-08-07   760   2011
#> 220 2020-08-08   222   1205
#> 221 2020-08-09   618   1837
#> 222 2020-08-10   145    932
#> 223 2020-08-11   667   1786
#> 224 2020-08-12   573   1614
#> 225 2020-08-13   574   1799
#> 226 2020-08-14   631   1998
#> 227 2020-08-15   947   2320
#> 228 2020-08-16   366   1606
#> 229 2020-08-17   214   1177
#> 230 2020-08-18   513   1566
#> 231 2020-08-19  1037   2382
#> 232 2020-08-20  1120   2895
#> 233 2020-08-21  1070   3230
#> 234 2020-08-22  1213   3729
#> 235 2020-08-23  1374   4214
#> 236 2020-08-24   752   3280
#> 237 2020-08-25   811   3171
#> 238 2020-08-26  2156   5292
#> 239 2020-08-27  1460   4685
#> 240 2020-08-28  1522   5106
#> 241 2020-08-29  1425   5018
#> 242 2020-08-30  1292   4768
#> 243 2020-08-31   729   3527
#> 244 2020-09-01   966   3596
#> 245 2020-09-02  1806   4875
#> 246 2020-09-03  1481   4719
#> 247 2020-09-04  2164   6189
#> 248 2020-09-05  1668   5751
#> 249 2020-09-06   996   4551
#> 250 2020-09-07   954   4014
#> 251 2020-09-08  1707   5056
#> 252 2020-09-09  1514   4893
#> 253 2020-09-10  1789   5607
#> 254 2020-09-11  1643   5572
#> 255 2020-09-12  1398   5219
#> 256 2020-09-13  1421   5145
#> 257 2020-09-14   702   3578
#> 258 2020-09-15  1509   4709
#> 259 2020-09-16  1730   5038
#> 260 2020-09-17  1744   5462
#> 261 2020-09-18  2306   6735
#> 262 2020-09-19  1423   5571
#> 263 2020-09-20  1550   5662
#> 264 2020-09-21  1159   4717
#> 265 2020-09-22  1448   5002
#> 266 2020-09-23  1958   5825
#> 267 2020-09-24  1970   6154
#> 268 2020-09-25  2068   6644
#> 269 2020-09-26  1841   6461
#> 270 2020-09-27  1682   6173
#> 271 2020-09-28  1274   5239
#> 272 2020-09-29  1845   5985
#> 273 2020-09-30  2101   6457
#> 274 2020-10-01  3550   9228
#> 275 2020-10-02  2471   8383
#> 276 2020-10-03  3270  10122
#> 277 2020-10-04  2350   8821
#> 278 2020-10-05  1987   7964
#> 279 2020-10-06  3212   9778
#> 280 2020-10-07  5109  13209
#> 281 2020-10-08  5433  15304
#> 282 2020-10-09  6512  18724
#> 283 2020-10-10  6130  19635
#> 284 2020-10-11  5222  18866
#> 285 2020-10-12  3937  16193
#> 286 2020-10-13  7603  21983
#> 287 2020-10-14  9174  25518
#> 288 2020-10-15 10690  30589
#> 289 2020-10-16 11450  34600
#> 290 2020-10-17 11978  37855
#> 291 2020-10-18 12618  40713
#> 292 2020-10-19  7495  32319
#> 293 2020-10-20 12782  40237
#> 294 2020-10-21 21492  54994
#> 295 2020-10-22 17161  54151
#> 296 2020-10-23 22918  67816
#> 297 2020-10-24 20316  67280
#> 298 2020-10-25 23186  74698
#> 299 2020-10-26 13694  58793
#> 300 2020-10-27 28806  83122
#> 301 2020-10-28 28661  85603
#> 302 2020-10-29 29057  92808
#> 303 2020-10-30 36278 109440
#> 304 2020-10-31 32758 109095
#> 305 2020-11-01 28348 103910
#> 306 2020-11-02 16687  78677
#> 307 2020-11-03 36423 107864
#> 308 2020-11-04 33439 104233
#> 309 2020-11-05 39546 120965
#> 310 2020-11-06 41970 131158
#> 311 2020-11-07 42393 138059
#> 312 2020-11-08 26944 112887
#> 313 2020-11-09 19795  90787
#> 314 2020-11-10 49888 136111
#> 315 2020-11-11 31512 110194
#> 316 2020-11-12 44617 136396
#> 317 2020-11-13 44696 141157
#> 318 2020-11-14 34389 128344
#> 319 2020-11-15 31450 119851
#> 320 2020-11-16 22351  96483
#> 321 2020-11-17 38921 119404
#> 322 2020-11-18 37381 117816
#> 323 2020-11-19 38914 125880
#> 324 2020-11-20 39093 129514
#> 325 2020-11-21 33069 120516
#> 326 2020-11-22 23474  99524
#> 327 2020-11-23 18975  82194
#> 328 2020-11-24 24445  84243
#> 329 2020-11-25 29690  91163
#> 330 2020-11-26 33603 101124
#> 331 2020-11-27 28527  97270
#> 332 2020-11-28 25043  91784
#> 333 2020-11-29 16535  72625
#> 334 2020-11-30 13425  59051
#> 335 2020-12-01 24011  72140
#> 336 2020-12-02 22973  71046
#> 337 2020-12-03 27356  81556
#> 338 2020-12-04 25912  82953
#> 339 2020-12-05 18908  72640
#> 340 2020-12-06 17409  66897
#> 341 2020-12-07 10252  48710
#> 342 2020-12-08 16831  55441
#> 343 2020-12-09 11343  43710
#> 344 2020-12-10 24343  64558
#> 345 2020-12-11 21595  63421
#> 346 2020-12-12 21910  68884
#> 347 2020-12-13 16568  61728
#> 348 2020-12-14  8294  43096
#> 349 2020-12-15 19699  58291
#> 350 2020-12-16 21797  60748
#> 351 2020-12-17 19688  62370
#> 352 2020-12-18 18430  62390
#> 353 2020-12-19 15249  56681
#> 354 2020-12-20 14296  53220
#> 355 2020-12-21  8046  38498
#> 356 2020-12-22 17311  51204
#> 357 2020-12-23 16518  49672
#> 358 2020-12-24 23215  64150
#> 359 2020-12-25 20611  64861
#> 360 2020-12-26  5796  38433
#> 361 2020-12-27  7838  34486
#> 362 2020-12-28  8702  30698
#> 363 2020-12-29 15891  41521
#> 364 2020-12-30 24647  58245
#> 365 2020-12-31 35222  83433
#> 366 2021-01-01 21433  73856
#> 367 2021-01-02  6406  44393
#> 368 2021-01-03 17667  58159
#> 369 2021-01-04  8392  36873
#> 370 2021-01-05 23293  60909
#> 371 2021-01-06 27869  70642
#> 372 2021-01-07 16302  60434
#> 373 2021-01-08 17665  62356
#> 374 2021-01-09 23406  71128
#> 375 2021-01-10 17685  63647
#> 376 2021-01-11  8638  44741
#> 377 2021-01-12 17019  54605
#> 378 2021-01-13 18113  54625
#> 379 2021-01-14 19498  59924
#> 380 2021-01-15 15544  55437
#> 381 2021-01-16 16973  57601
#> 382 2021-01-17  9876  43658
#> 383 2021-01-18  6399  32134
#> 384 2021-01-19 13399  40396
#> 385 2021-01-20 18464  48045
#> 386 2021-01-21 15185  47603
#> 387 2021-01-22 13658  47204
#> 388 2021-01-23 13518  46691
#> 389 2021-01-24 10375  40441