Estimate the proportion of asymptomatic cases by capture/recapture

Estimate the proportion of asymptomatic cases by capture/recapture

estimate_asympto(df, bounds = c("lower", "upper"))

Arguments

df
A data.frame containing three columns containing date, the daily number of new_cases, and the daily number of new_deaths.

bounds
"lower", "upper", or both c("lower, "upper") (the default), telling which bounds of the number of asymptomatic cases are computed. associated with the lower and upper bounds. estimation. FALSE is never the recommended setting excepted for testing or debugging purpose.

Value

A data.frame with two or three columns (depending on the value of the bounds argument):

date: the original date column from your input dataset df

lower: the lower bound of asymptomatic cases

upper: the upper bound of asymptomatic cases

References

Böhning D., Rocchetti I., Maruotti A., Holling H. (2020), Estimating the undetected infections in the Covid-19 outbreak by harnessing capture–recapture methods, International Journal of Infectious Diseases, 97, p197-201, doi: 10.1016/j.ijid.2020.06.009 .

Rocchetti I., Böhning D., Holling H., Maruotti A., (2020), Estimating the size of undetected cases of the SARS-CoV-2 outbreak in Europe: An upperbound estimator, medRxiv, doi: 10.1101/2020.07.14.20153445 .

Examples

d <- readRDS(system.file("extdata", "covid19_italy.rds", package = "asymptor"))
head(d)

#>         date new_cases new_deaths
#> 1 2019-12-31         0          0
#> 2 2020-01-01         0          0
#> 3 2020-01-02         0          0
#> 4 2020-01-03         0          0
#> 5 2020-01-04         0          0
#> 6 2020-01-05         0          0

estimate_asympto(d)

#>           date lower upper
#> 1   2019-12-31    NA    NA
#> 2   2020-01-01     0    NA
#> 3   2020-01-02     0    NA
#> 4   2020-01-03     0    NA
#> 5   2020-01-04     0    NA
#> 6   2020-01-05     0    NA
#> 7   2020-01-06     0    NA
#> 8   2020-01-07     0    NA
#> 9   2020-01-08     0    NA
#> 10  2020-01-09     0    NA
#> 11  2020-01-10     0    NA
#> 12  2020-01-11     0    NA
#> 13  2020-01-12     0    NA
#> 14  2020-01-13     0    NA
#> 15  2020-01-14     0    NA
#> 16  2020-01-15     0    NA
#> 17  2020-01-16     0    NA
#> 18  2020-01-17     0    NA
#> 19  2020-01-18     0    NA
#> 20  2020-01-19     0    NA
#> 21  2020-01-20     0    NA
#> 22  2020-01-21     0    NA
#> 23  2020-01-22     0    NA
#> 24  2020-01-23     0    NA
#> 25  2020-01-24     0    NA
#> 26  2020-01-25     0    NA
#> 27  2020-01-26     0    NA
#> 28  2020-01-27     0    NA
#> 29  2020-01-28     0    NA
#> 30  2020-01-29     2    NA
#> 31  2020-01-30     0     2
#> 32  2020-01-31     0    NA
#> 33  2020-02-01     0     1
#> 34  2020-02-02     0    NA
#> 35  2020-02-03     0    NA
#> 36  2020-02-04     0    NA
#> 37  2020-02-05     0    NA
#> 38  2020-02-06     0    NA
#> 39  2020-02-07     0    NA
#> 40  2020-02-08     0    NA
#> 41  2020-02-09     0    NA
#> 42  2020-02-10     0    NA
#> 43  2020-02-11     0    NA
#> 44  2020-02-12     0    NA
#> 45  2020-02-13     0    NA
#> 46  2020-02-14     0    NA
#> 47  2020-02-15     0    NA
#> 48  2020-02-16     0    NA
#> 49  2020-02-17     0    NA
#> 50  2020-02-18     0    NA
#> 51  2020-02-19     0    NA
#> 52  2020-02-20     0    NA
#> 53  2020-02-21     0    NA
#> 54  2020-02-22   182    NA
#> 55  2020-02-23   291   357
#> 56  2020-02-24   772  1087
#> 57  2020-02-25    39   357
#> 58  2020-02-26    65   318
#> 59  2020-02-27   841  1388
#> 60  2020-02-28   228   745
#> 61  2020-02-29   248   844
#> 62  2020-03-01  1355  2537
#> 63  2020-03-02   212  1154
#> 64  2020-03-03   686  1924
#> 65  2020-03-04   784  2028
#> 66  2020-03-05  1080  2721
#> 67  2020-03-06   838  2615
#> 68  2020-03-07  2091  4799
#> 69  2020-03-08  1995  5067
#> 70  2020-03-09  2312  6240
#> 71  2020-03-10   585  3466
#> 72  2020-03-11  6838 13100
#> 73  2020-03-12  3306  8491
#> 74  2020-03-13  2700  8670
#> 75  2020-03-14  5152 12975
#> 76  2020-03-15  4116 12063
#> 77  2020-03-16  3223 11081
#> 78  2020-03-17  4302 12659
#> 79  2020-03-18  5798 14983
#> 80  2020-03-19  7490 18751
#> 81  2020-03-20  7629 20562
#> 82  2020-03-21  8276 22775
#> 83  2020-03-22  5232 18771
#> 84  2020-03-23  4623 16887
#> 85  2020-03-24  6807 19189
#> 86  2020-03-25  5942 17763
#> 87  2020-03-26  8321 22058
#> 88  2020-03-27  6847 20190
#> 89  2020-03-28  7037 20817
#> 90  2020-03-29  5214 17805
#> 91  2020-03-30  3722 14116
#> 92  2020-03-31  5110 14885
#> 93  2020-04-01  6872 17087
#> 94  2020-04-02  5415 15704
#> 95  2020-04-03  5385 15938
#> 96  2020-04-04  5911 16806
#> 97  2020-04-05  4350 14656
#> 98  2020-04-06  3518 12558
#> 99  2020-04-07  3082 10687
#> 100 2020-04-08  5889 14503
#> 101 2020-04-09  5475 14296
#> 102 2020-04-10  4293 13382
#> 103 2020-04-11  6609 17118
#> 104 2020-04-12  3926 13644
#> 105 2020-04-13  2818 11056
#> 106 2020-04-14  3460 10815
#> 107 2020-04-15  2969  9187
#> 108 2020-04-16  6687 14723
#> 109 2020-04-17  3797 11389
#> 110 2020-04-18  4045 12245
#> 111 2020-04-19  3034 10420
#> 112 2020-04-20  1961  7863
#> 113 2020-04-21  4321 10771
#> 114 2020-04-22  4951 11737
#> 115 2020-04-23  2408  8752
#> 116 2020-04-24  4097 11285
#> 117 2020-04-25  2130  7870
#> 118 2020-04-26  2573  8360
#> 119 2020-04-27  1517  5965
#> 120 2020-04-28  3218  8168
#> 121 2020-04-29  2459  6972
#> 122 2020-04-30  1944  6401
#> 123 2020-05-01  2406  7007
#> 124 2020-05-02  2418  6614
#> 125 2020-05-03  1116  4703
#> 126 2020-05-04  1247  4452
#> 127 2020-05-05  1171  3775
#> 128 2020-05-06  2947  5942
#> 129 2020-05-07  1675  4527
#> 130 2020-05-08  1518  4570
#> 131 2020-05-09  1033  3699
#> 132 2020-05-10   699  2812
#> 133 2020-05-11   886  2732
#> 134 2020-05-12  3428  6185
#> 135 2020-05-13   652  2864
#> 136 2020-05-14  1568  4102
#> 137 2020-05-15   828  2573
#> 138 2020-05-16  1201  3186
#> 139 2020-05-17   622  2245
#> 140 2020-05-18   352  1595
#> 141 2020-05-19  2276  4234
#> 142 2020-05-20   676  2059
#> 143 2020-05-21   807  2306
#> 144 2020-05-22   827  2261
#> 145 2020-05-23   837  2312
#> 146 2020-05-24   454  1795
#> 147 2020-05-25   204  1088
#> 148 2020-05-26   705  1735
#> 149 2020-05-27  1212  2223
#> 150 2020-05-28   682  1918
#> 151 2020-05-29   524  1765
#> 152 2020-05-30   425  1451
#> 153 2020-05-31   367  1238
#> 154 2020-06-01   106   643
#> 155 2020-06-02   813  1696
#> 156 2020-06-03   414  1020
#> 157 2020-06-04   133   595
#> 158 2020-06-05  2880  4188
#> 159 2020-06-06   162   883
#> 160 2020-06-07   177   759
#> 161 2020-06-08   587  1216
#> 162 2020-06-09   395   922
#> 163 2020-06-10   191   664
#> 164 2020-06-11   955  1771
#> 165 2020-06-12    82   583
#> 166 2020-06-13  1388  2608
#> 167 2020-06-14   376  1040
#> 168 2020-06-15   292  1039
#> 169 2020-06-16   163   736
#> 170 2020-06-17   642  1431
#> 171 2020-06-18   419  1086
#> 172 2020-06-19   219   844
#> 173 2020-06-20   337   973
#> 174 2020-06-21   209   752
#> 175 2020-06-22   234   773
#> 176 2020-06-23    73   437
#> 177 2020-06-24   233   733
#> 178 2020-06-25   556  1155
#> 179 2020-06-26   250   835
#> 180 2020-06-27   121   620
#> 181 2020-06-28   195   659
#> 182 2020-06-29    93   434
#> 183 2020-06-30   193   551
#> 184 2020-07-01   285   669
#> 185 2020-07-02   254   677
#> 186 2020-07-03   265   766
#> 187 2020-07-04   271   810
#> 188 2020-07-05   160   654
#> 189 2020-07-06   233   755
#> 190 2020-07-07   106   473
#> 191 2020-07-08   299   769
#> 192 2020-07-09   287   775
#> 193 2020-07-10   348   961
#> 194 2020-07-11   130   647
#> 195 2020-07-12   303   907
#> 196 2020-07-13   128   572
#> 197 2020-07-14    84   415
#> 198 2020-07-15   259   659
#> 199 2020-07-16   366   818
#> 200 2020-07-17   246   769
#> 201 2020-07-18   281   872
#> 202 2020-07-19   193   748
#> 203 2020-07-20   173   672
#> 204 2020-07-21    94   457
#> 205 2020-07-22   655  1327
#> 206 2020-07-23   342   985
#> 207 2020-07-24   209   864
#> 208 2020-07-25   304   998
#> 209 2020-07-26   239   874
#> 210 2020-07-27   114   607
#> 211 2020-07-28   280   831
#> 212 2020-07-29   402  1021
#> 213 2020-07-30   518  1346
#> 214 2020-07-31   379  1274
#> 215 2020-08-01   231  1029
#> 216 2020-08-02   198   861
#> 217 2020-08-03   110   575
#> 218 2020-08-04   232   725
#> 219 2020-08-05   813  1584
#> 220 2020-08-06   425  1300
#> 221 2020-08-07   760  2011
#> 222 2020-08-08   222  1205
#> 223 2020-08-09   618  1837
#> 224 2020-08-10   145   932
#> 225 2020-08-11   667  1786
#> 226 2020-08-12   573  1614
#> 227 2020-08-13   574  1799
#> 228 2020-08-14   631  1998
#> 229 2020-08-15   947  2320
#> 230 2020-08-16   366  1606
#> 231 2020-08-17   214  1177
#> 232 2020-08-18   513  1566
#> 233 2020-08-19  1037  2382
#> 234 2020-08-20  1120  2895
#> 235 2020-08-21  1070  3230
#> 236 2020-08-22  1213  3729
#> 237 2020-08-23  1374  4214
#> 238 2020-08-24   752  3280
#> 239 2020-08-25   811  3171
#> 240 2020-08-26  2156  5292
#> 241 2020-08-27  1460  4685
#> 242 2020-08-28  1522  5106
#> 243 2020-08-29  1425  5018
#> 244 2020-08-30  1292  4768
#> 245 2020-08-31   729  3527
#> 246 2020-09-01   966  3596
#> 247 2020-09-02  1806  4875
#> 248 2020-09-03  1481  4719
#> 249 2020-09-04  2164  6189
#> 250 2020-09-05  1668  5751
#> 251 2020-09-06   996  4551
#> 252 2020-09-07   954  4014
#> 253 2020-09-08  1707  5056
#> 254 2020-09-09  1514  4893
#> 255 2020-09-10  1789  5607
#> 256 2020-09-11  1643  5572
#> 257 2020-09-12  1398  5219
#> 258 2020-09-13  1421  5145
#> 259 2020-09-14   702  3578
#> 260 2020-09-15  1509  4709
#> 261 2020-09-16  1730  5038
#> 262 2020-09-17  1744  5462
#> 263 2020-09-18  2306  6735
#> 264 2020-09-19  1423  5571
#> 265 2020-09-20  1550  5662
#> 266 2020-09-21  1159  4717
#> 267 2020-09-22  1448  5002
#> 268 2020-09-23  1958  5825
#> 269 2020-09-24  1970  6154
#> 270 2020-09-25  2068  6644
#> 271 2020-09-26  1841  6461
#> 272 2020-09-27  1682  6173
#> 273 2020-09-28  1274  5239
#> 274 2020-09-29  1845  5985
#> 275 2020-09-30  2101  6457
#> 276 2020-10-01  3550  9228
#> 277 2020-10-02  2471  8383
#> 278 2020-10-03  3270 10122
#> 279 2020-10-04  2350  8821
#> 280 2020-10-05  1987  7964
#> 281 2020-10-06  3212  9778
#> 282 2020-10-07  5109 13209
#> 283 2020-10-08  5433 15304
#> 284 2020-10-09  6512 18724
#> 285 2020-10-10  6130 19635
#> 286 2020-10-11  5222 18866
#> 287 2020-10-12  3937 16193
#> 288 2020-10-13  7603 21983
#> 289 2020-10-14  9174 25518
#> 290 2020-10-15 10690 30589
#> 291 2020-10-16 11450 34600
#> 292 2020-10-17 11978 37855
#> 293 2020-10-18 12618 40713
#> 294 2020-10-19  7495 32319
#> 295 2020-10-20 12782 40237
#> 296 2020-10-21 21492 54994