Estimate the proportion of asymptomatic cases by capture/recapture

Estimate the proportion of asymptomatic cases by capture/recapture

estimate_asympto(df, bounds = c("lower", "upper"))

Arguments

df
A data.frame containing three columns containing date, the daily number of cases, and the daily number of deaths. To compute the upper bound, an additional recoveries column is required

bounds
"lower", "upper", or both c("lower, "upper") (the default), telling which bounds of the number of asymptomatic cases are computed. associated with the lower and upper bounds. estimation. FALSE is never the recommended setting excepted for testing or debugging purpose.

Value

A data.frame with two or three columns (depending on the value of the bounds argument):

date: the original date column from your input dataset df

lower: the lower bound of asymptomatic cases

upper: the upper bound of asymptomatic cases

References

Böhning D., Rocchetti I., Maruotti A., Holling H. (2020), Estimating the undetected infections in the Covid-19 outbreak by harnessing capture–recapture methods, International Journal of Infectious Diseases, 97, p197-201, doi: 10.1016/j.ijid.2020.06.009 .

Rocchetti I., Böhning D., Holling H., Maruotti A., (2020), Estimating the size of undetected cases of the SARS-CoV-2 outbreak in Europe: An upperbound estimator, medRxiv, doi: 10.1101/2020.07.14.20153445 .